Move the definitions of subparsers into the modules for subcommands
