+++ onlinejudge_command/subcommand/submit.py
@@ -16,6 +16,38 @@ from onlinejudge.type import *
 logger = getLogger(__name__)
 
 
+def add_subparser(subparsers: argparse.Action) -> None:
+    subparsers_add_parser: Callable[..., argparse.ArgumentParser] = subparsers.add_parser  # type: ignore
+    subparser = subparsers_add_parser('submit', aliases=['s'], help='submit your solution', formatter_class=argparse.RawTextHelpFormatter, epilog='''\
+supported services:
+  AtCoder
+  Codeforces
+  yukicoder
+  HackerRank
+  Toph (Problem Archive)
+
+tips:
+  This subcommand has the feature to guess the problem to submit to. To guess the problem, run `oj download https://...` in the same directory without `--directory` option before using `oj submit ...`.
+
+  you can do similar things with shell and oj-api command. see https://github.com/online-judge-tools/api-client
+    e.g. $ oj-api submit-code --file main.cpp --language $(oj-api guess-language-id --file main.cpp https://atcoder.jp/contests/agc001/tasks/agc001_a | jq -r .result.id) https://atcoder.jp/contests/agc001/tasks/agc001_a
+''')
+    subparser.add_argument('url', nargs='?', help='the URL of the problem to submit. if not given, guessed from history of download command.')
+    subparser.add_argument('file', type=pathlib.Path)
+    subparser.add_argument('-l', '--language', help='narrow down language choices if ambiguous')
+    subparser.add_argument('--no-guess', action='store_false', dest='guess')
+    subparser.add_argument('-g', '--guess', action='store_true', help='guess the language for your file (default)')
+    subparser.add_argument('--no-guess-latest', action='store_false', dest='guess_cxx_latest')
+    subparser.add_argument('--guess-cxx-latest', action='store_true', help='use the lasest version for C++ (default)')
+    subparser.add_argument('--guess-cxx-compiler', choices=('gcc', 'clang', 'all'), default='gcc', help='use the specified C++ compiler if both of GCC and Clang are available (default: gcc)')
+    subparser.add_argument('--guess-python-version', choices=('2', '3', 'auto', 'all'), default='auto', help='default: auto')
+    subparser.add_argument('--guess-python-interpreter', choices=('cpython', 'pypy', 'all'), default='cpython', help='use the specified Python interpreter if both of CPython and PyPy are available (default: cpython)')
+    subparser.add_argument('--no-open', action='store_false', dest='open')
+    subparser.add_argument('--open', action='store_true', default=True, help='open the result page after submission (default)')
+    subparser.add_argument('-w', '--wait', metavar='SECOND', type=float, default=3, help='sleep before submitting')
+    subparser.add_argument('-y', '--yes', action='store_true', help='don\'t confirm')
+
+
 def submit(args: argparse.Namespace) -> None:
     # guess url
     history = onlinejudge_command.download_history.DownloadHistory()
